src/requirements.txt
pandas
pyyaml
rpaframework
